+++ CHANGELOG.rst
@@ -9,6 +9,10 @@ Unreleased - YYYY-MM-DD
 
 Features and Enhancements
 -------------------------
+- Add leader status to the ``synapse.tools.aha.list`` tool output.
+  This will only be available if a leader has been registered for
+  the service.
+  (`#3078 <https://github.com/vertexproject/synapse/pull/3078>`_)
 - Add support for private values in Storm modules, which are specified
   by beginning the name with a double underscore (``__``). These values
   cannot be dereferenced outside of the module they are declared in.